rename extension, fix refactor errors, fix readme configs
