+++ service/defaultcomponents/defaults.go
@@ -30,7 +30,7 @@ import (
 	"go.opentelemetry.io/collector/exporter/zipkinexporter"
 	"go.opentelemetry.io/collector/extension/fluentbitextension"
 	"go.opentelemetry.io/collector/extension/healthcheckextension"
-	"go.opentelemetry.io/collector/extension/oidcextension"
+	"go.opentelemetry.io/collector/extension/authoidcextension"
 	"go.opentelemetry.io/collector/extension/pprofextension"
 	"go.opentelemetry.io/collector/extension/zpagesextension"
 	"go.opentelemetry.io/collector/processor/attributesprocessor"
@@ -64,7 +64,7 @@ func Components() (
 		pprofextension.NewFactory(),
 		zpagesextension.NewFactory(),
 		fluentbitextension.NewFactory(),
-		oidcextension.NewFactory(),
+		authoidcextension.NewFactory(),
 	)
 	if err != nil {
 		errs = append(errs, err)
